Please comply with the terms of the license
Please remove https://play.google.com/store/apps/details?id=de.alkhemeia.suntimeswidget from Google Play - it isn't compliant with the license.

Over the last week I have attempted to get Google to remove this for me. I would like to apologize for that. I should have contacted you first before jumping straight to action.
Google has decided not to act at this time and instead wait for us to resolve this between ourselves. In retrospect I think this is fortunate, because it will allow you to remove the application without being penalized by Google, or alternatively, fix your fork so it is in full compliance with the GPL. I would really prefer that you simply remove it though - it does not add anything but advertising and reflects poorly on the project. Furthermore its unethical! If you could see the history/balance on the PayPal account I use to collect donations, you might think twice from sticking your hand in the jar. :(

I really hope we can resolve this amicably. I have invested a lot of time in this project and I'm not going away.